ValueInput refactroed to its own file
